Adjust rendering integration test to account for new greelist
