+++ contracts/test/GatedMint.spec.ts
@@ -1,10 +1,11 @@
 import { expect } from "chai";
-import hre, { deployments, ethers } from "hardhat";
+import hre, { deployments } from "hardhat";
 import "@nomiclabs/hardhat-ethers";
 
 import { MerkleTree } from "merkletreejs";
 import { keccak256 } from "ethers/lib/utils";
-import { BigNumber, Signer, Wallet } from "ethers";
+
+import { getPermit, getPermitSigners } from "./createMintPermit";
 
 describe("GatedMint", async () => {
   const baseSetup = deployments.createFixture(async () => {
@@ -77,46 +78,3 @@ describe("GatedMint", async () => {
     });
   });
 });
-
-function getPermitSigners() {
-  const passwords = [
-    "double custom typical style harvest buzz",
-    "wasp decorate badge coast park pole",
-    "sense pudding cause leave security subway",
-  ];
-
-  return passwords
-    .map((password) =>
-      ethers.utils.keccak256(ethers.utils.toUtf8Bytes(password))
-    )
-    .map((pk) => new ethers.Wallet(pk));
-}
-
-async function getPermit(from: number, to: number) {
-  const permitSigners = getPermitSigners();
-  const elements = await Promise.all(permitSigners.map((s) => s.getAddress()));
-  const merkleTree = new MerkleTree(elements, keccak256, {
-    hashLeaves: true,
-    sortPairs: true,
-  });
-
-  const permitSigner = permitSigners[from] as Signer;
-  const permitSignerAddress = await permitSigner.getAddress();
-
-  const minterSigner = (await hre.ethers.getSigners())[to];
-  const minterAddress = await minterSigner.getAddress();
-
-  // not we have to arrayify because wallet.signMessage does not recognize hex strings
-  const messageHash = ethers.utils.arrayify(keccak256(minterAddress));
-  const signature = await permitSigner.signMessage(messageHash);
-
-  expect(ethers.utils.verifyMessage(messageHash, signature)).to.eq(
-    permitSignerAddress
-  );
-
-  return {
-    signature,
-    signer: permitSignerAddress,
-    proof: merkleTree.getHexProof(keccak256(permitSignerAddress)),
-  };
-}
